[commom] Add auto installation script
- Add install script 'install_vim.sh'
- Modify 'README.md' to fit the script
- Modify script 'install_ccls.sh' for easier usage
- Remove 'INSTALL_CCLS.md' due to 'install_ccls.sh'

+++ setup/install_global.sh
@@ -5,28 +5,32 @@
 set -e
 set -v
 
+pushd $(dirname ${BASH_SOURCE[0]}) > /dev/null
+
 VERSION="6.6.3"
 TARFILE=global-$VERSION.tar.gz
 DIR=global-$VERSION
 
-sudo apt-get update
-sudo apt-get install -y \
-    autoconf \
-    gperf \
-    bison \
-    flex \
-    libtool-bin \
-    libncurses-dev \
-    texinfo
+if [ ! -d $DIR ]; then
+    sudo apt-get update
+    sudo apt-get install -y \
+        gperf \
+        bison \
+        flex \
+        libtool-bin \
+        libncurses-dev \
+        texinfo
 
-wget http://tamacom.com/global/$TARFILE
+    wget http://tamacom.com/global/$TARFILE
+    tar -xvf $TARFILE > /dev/null
+    rm $TARFILE
+fi
 
-tar -xvf $TARFILE > /dev/null
-rm $TARFILE
 pushd $DIR > /dev/null
 sh reconf.sh
 ./configure
 make -j
 sudo make install
 
+popd > /dev/null
 popd > /dev/null
